extract logic back into presenter
only return posts from repository, don't do logic there
